figma-transformer: extract vector geometry normalizer
